+++ android/src/main/kotlin/ru/kovardin/flutter_yandex_ads/components/YandexAdsInterstitial.kt
@@ -26,7 +26,7 @@ class YandexAdsInterstitial(private val api: YandexApi, private val context: Con
                     val builder = Yandex.EventResponse.Builder()
 
                     val response = builder.build()
-                    api.callbacks.get(EventKey(id = id
+                    api.callbacks.remove(EventKey(id = id
                             ?: "", name = "onAdLoaded", type = EventType.INTERSTITIAL.type))?.success(response)
                 }
 
@@ -36,7 +36,7 @@ class YandexAdsInterstitial(private val api: YandexApi, private val context: Con
                     builder.setDescription(error.description)
 
                     val response = builder.build()
-                    api.callbacks.get(EventKey(id = id
+                    api.callbacks.remove(EventKey(id = id
                             ?: "", name = "onAdFailedToLoad", type = EventType.INTERSTITIAL.type))?.success(response)
                 }
 
@@ -46,14 +46,14 @@ class YandexAdsInterstitial(private val api: YandexApi, private val context: Con
                     val response = builder.build()
                     val key = EventKey(id = id
                             ?: "", name = "onAdShown", type = EventType.INTERSTITIAL.type)
-                    api.callbacks.get(key)?.success(response)
+                    api.callbacks.remove(key)?.success(response)
                 }
 
                 override fun onAdDismissed() {
                     val builder = Yandex.EventResponse.Builder()
 
                     val response = builder.build()
-                    api.callbacks.get(EventKey(id = id
+                    api.callbacks.remove(EventKey(id = id
                             ?: "", name = "onAdDismissed", type = EventType.INTERSTITIAL.type))?.success(response)
                 }
 
@@ -61,7 +61,7 @@ class YandexAdsInterstitial(private val api: YandexApi, private val context: Con
                     val builder = Yandex.EventResponse.Builder()
 
                     val response = builder.build()
-                    api.callbacks.get(EventKey(id = id
+                    api.callbacks.remove(EventKey(id = id
                             ?: "", name = "onAdClicked", type = EventType.INTERSTITIAL.type))?.success(response)
                 }
 
@@ -69,7 +69,7 @@ class YandexAdsInterstitial(private val api: YandexApi, private val context: Con
                     val builder = Yandex.EventResponse.Builder()
 
                     val response = builder.build()
-                    api.callbacks.get(EventKey(id = id
+                    api.callbacks.remove(EventKey(id = id
                             ?: "", name = "onLeftApplication", type = EventType.INTERSTITIAL.type))?.success(response)
                 }
 
@@ -77,7 +77,7 @@ class YandexAdsInterstitial(private val api: YandexApi, private val context: Con
                     val builder = Yandex.EventResponse.Builder()
 
                     val response = builder.build()
-                    api.callbacks.get(EventKey(id = id
+                    api.callbacks.remove(EventKey(id = id
                             ?: "", name = "onReturnedToApplication", type = EventType.INTERSTITIAL.type))?.success(response)
                 }
 
@@ -86,7 +86,7 @@ class YandexAdsInterstitial(private val api: YandexApi, private val context: Con
                     builder.setData(data?.rawData ?: "")
 
                     val response = builder.build()
-                    api.callbacks.get(EventKey(id = id
+                    api.callbacks.remove(EventKey(id = id
                             ?: "", name = "onImpression", type = EventType.INTERSTITIAL.type))?.success(response)
                 }
             })