+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Fixed exception on ad callbacks
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Banner and Interstitial formats for Android and iOS